+++ kubejs/server_scripts/tags/fluids/neoforge/eclipse_ember_fuel.js
@@ -1,0 +1,3 @@
+ServerEvents.tags('fluid', (event) => {
+    event.get('c:fuels/eclipse_ember').add(['justdirethings:refined_t4_fluid_source']);
+});