+++ kubejs/server_scripts/tags/fluids/neoforge/voidflame_fuel.js
@@ -1,0 +1,3 @@
+ServerEvents.tags('fluid', (event) => {
+    event.get('c:fuels/voidflame').add(['justdirethings:refined_t3_fluid_source']);
+});